+++ src/taoensso/nippy.clj
@@ -68,6 +68,7 @@
   (def ^:const id-integer    (int 42))
   (def ^:const id-long       (int 43))
   (def ^:const id-bigint     (int 44))
+  (def ^:const id-biginteger (int 45))
 
   (def ^:const id-float      (int 60))
   (def ^:const id-double     (int 61))
@@ -246,8 +247,8 @@
 
 ;;
 
-(freezer BigInt     id-bigint  (write-biginteger out (.toBigInteger x)))
-(freezer BigInteger id-bigint  (write-biginteger out x))
+(freezer BigInt     id-bigint     (write-biginteger out (.toBigInteger x)))
+(freezer BigInteger id-biginteger (write-biginteger out x))
 
 (freezer Float      id-float   (.writeFloat  out x))
 (freezer Double     id-double  (.writeDouble out x))
@@ -423,7 +424,8 @@
         id-int-as-long   (long (.readInt   in))
         ;; id-compact-long  (read-compact-long in)
 
-        id-bigint  (bigint (read-biginteger in))
+        id-bigint     (bigint (read-biginteger in))
+        id-biginteger (read-biginteger in)
 
         id-float  (.readFloat  in)
         id-double (.readDouble in)
@@ -496,6 +498,10 @@
                  :as   opts}]]
 
   (let [headerless-meta (merge headerless-meta (:legacy-opts opts)) ; Deprecated
+        _ (assert (or (nil? headerless-meta)
+                      (head-meta-id headerless-meta))
+                  "Bad :headerless-meta (should be nil or a valid `head-meta` value)")
+
         ex (fn [msg & [e]] (throw (Exception. (str "Thaw failed: " msg) e)))
         try-thaw-data
         (fn [data-ba {:keys [compressed? encrypted?] :as _head-or-headerless-meta}]
@@ -593,8 +599,7 @@
 
 (defrecord Compressable-LZMA2 [value])
 (extend-freeze Compressable-LZMA2 128 [x out]
-  (let [[_ ^bytes ba] (-> (freeze (:value x) {:compressor nil})
-                          (utils/ba-split 4))
+  (let [ba (freeze (:value x) {:skip-header? true :compressor nil})
         ba-len    (alength ba)
         compress? (> ba-len 1024)]
     (.writeBoolean out compress?)
@@ -605,8 +610,10 @@
 (extend-thaw 128 [in]
   (let [compressed? (.readBoolean in)
         ba          (read-bytes in)]
-    (thaw (wrap-header ba {:compressed? compressed? :encrypted? false})
-          {:compressor compression/lzma2-compressor})))
+    (thaw ba {:compressor compression/lzma2-compressor
+              :headerless-meta {:version     1
+                                :compressed? compressed?
+                                :encrypted?  false}})))
 
 (comment
   (->> (apply str (repeatedly 1000 rand))
@@ -690,7 +697,9 @@
   (dissoc stress-data :bytes :throwable :exception :ex-info :queue :queue-empty
                       :byte :stress-record))
 
-;;;; Data recovery/analysis
+;;;; Tools
+
+(utils/defalias freezeable? utils/freezable?)
 
 (defn inspect-ba "Alpha - subject to change."
   [ba & [thaw-opts]]