+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v2.6.0-alpha2 / 2014-Jan-23
+## v2.6.0-alpha4 / 2014-Feb-16
 
 **WARNING**: This is an **EXPERIMENTAL early testing release** and **unsuitable for use in production**. Welcoming feedback on any issues, etc.!
 
@@ -8,12 +8,17 @@
  * New test suite added to ensure a 1-to-1 value->binary representation mapping for all core data types. This will be a guarantee kept going forward.
  * New `:skip-header?` `freeze` option to freeze data without standard Nippy headers (can be useful in very performance sensitive environments).
  * New benchmarks added, notably a Fressian comparison.
+ * Added experimental `freezable?` util fn to main ns.
+ * Added some property-based [simple-check](https://github.com/reiddraper/simple-check) roundtrip tests.
+
 
 ### Changes
  * **BREAKING**: the experimental `Compressable-LZMA2` type has changed (less overhead).
  * **DEPRECATED**: `freeze-to-stream!`, `thaw-from-stream!` are deprecated in favor of the more general `freeze-to-out!`, `thaw-from-in!`.
  * **DEPRECATED**: `:legacy-mode` options. This was being used mainly for headerless freezing, so a new headerless mode is taking its place.
  * Public utils now available for custom type extension: `write-bytes`, `write-biginteger`, `write-utf8`, `write-compact-long`, and respective readers.
+ * Now distinguish between `BigInteger` and `BigInt` on thawing (previously both thawed to `BigInt`s). (mlacorte).
+
 
 ### Fixes
  * None.